[WIP]: rockchip/mpp_device: common device design
--------------    ---------------    ------------------
|            |    |             |    | mpp device     |
| mpp device |<-->| mpp service |<-->|  implementation|
|    top     |    |             |    |  and buttom    |
--------------    ---------------    ------------------

The mpp common device will do the the top part, but still
allow the mpp device implementation change the default behaviour.

I still can't remove the full mpp service from mpp rkdev device,
as the link mode may need it, it is disabled now.

The mpp device iplementation should only register its to the
mpp device framework. But the mpp common device is the base
class not the top class.

The link mode will be regard as a bundle of tasks.

The functions with rockchip_mpp_* prefix are used
internal for the task operation.

This device mode is not just suitable for the video IP but
also IEP I think.

I know the mpp common device become a module and
a different mpp device implementation will be module as well.

Change-Id: I97dd4eaf20a925f820133a6cb2d954fb97367cb3
Signed-off-by: Randy Li <randy.li@rock-chips.com>